Base project to parse SQL queries and visualize
This is first commit. Starting with parsing SQL/OLAP queries and extract
information required for Cassandra data modeling.
This code parses SQL queries and produces JSON which is used to
visualize large complex queries. A separate project queryVis is added to
visualize.
Generate JSON by executing main program in modeler.SQLParserUtil and
paste printed JSON in queryViz's index.html and open in browser. modeler
is an eclipse project, so can be imported in eclipde.

This is useful in visualize large query for understanding easily. This
helps in debugging as seen in example in queryViz table t3 is not
connected, hence no join condition to match, causing unexpected result.

Limitations:
All queries must have column names prefixed with table names and no '*'
or column position like "select *" or "select 1" or "select max(1)".
This is a TODO and will be done in future.

Subqueries not completely tested and queryVis does not have support for
this. This is also a TODO and will be done.